added modular models
