+++ src/models/index.js
@@ -1,0 +1,56 @@
+let users = {
+    1: {
+      id: '1',
+      username: 'Conrad P.B',
+    },
+    2: {
+      id: '2',
+      username: 'Tia reed',
+    },
+    3: {
+      id: '3',
+      username: 'Sofi trey',
+    },
+  }
+  
+let questions = {
+    1: {
+      id: '1',
+      text: 'Hello all, How u today?',
+      userId: '1',
+    },
+    2: {
+      id: '2',
+      text: 'Whats the weather like outside?',
+      userId: '2',
+    },
+    3: {
+      id: '3',
+      text: 'So, shall we go?',
+      userId: '3',
+    },
+  }
+
+let answers = {
+    1: {
+      id: '1',
+      text: 'Hi, we gud',
+      userId: '1',
+    },
+    2: {
+      id: '2',
+      text: 'Itsss all gooodd brahh, weathers fine!!',
+      userId: '2',
+    },
+    3: {
+      id: '3',
+      text: 'Ya.. Lets go.. will be dark soon',
+      userId: '3',
+    },
+  }
+
+  export default {
+    users,
+    questions,
+    answers
+  }
